Refactor code using ES6 syntaxe
