+++ index.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="./script/script.js" defer type="module"></script>
-    <link rel="stylesheet" href="style.css">
+    <script src="index.js" defer type="module"></script>
+    <link rel="stylesheet" href="index.css">
     <title>Awesome Books</title>
   </head>
