+++ modules/displayNoBook.js
@@ -1,0 +1,9 @@
+const noBook = document.querySelector('.no-book');
+
+export default (bookLength, container) => {
+  noBook.style.setProperty('display', `${bookLength ? 'none' : 'block'}`);
+  container.style.setProperty(
+    'display',
+    `${!bookLength ? 'none' : 'block'}`,
+  );
+};
