+++ modules/saveLocally.js
@@ -1,0 +1,3 @@
+export default (data) => {
+  localStorage.setItem('books', JSON.stringify(data));
+};
